Added ASIHTTPRequest and JSONKit
Warning! This project uses ARC and ASIHTTPRequest and JSONKit do not.
This can cause potential instability since the behavior for mixing ARC and non-ARC for these two projects is undefined.

To get the project to compile non-ARC files are being compiled with the -fno-objc-arc flag, since ARC is compiled on a per file basis.